pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "moatless-tree-search"
version = "0.0.4"
description = ""
authors = [ "Albert Örwall <albert@moatless.ai>", "Antonis Antoniades <antonis@ucsb.edu>"]
readme = "README.md"
packages = [
    { include = "moatless" },
]

[tool.poetry.scripts]
moatless-streamlit = "moatless.streamlit.cli:main_run"
moatless-evaluate = "moatless.benchmark.run_evaluation:main"

[tool.ruff]
target-version = "py310"
extend-exclude = [ "tests", "evaluations", "notebooks",]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"

pydantic = "^2.8.2"
tiktoken = "^0.7.0"
networkx = "^3.3"
tree-sitter = "0.22.3"
tree-sitter-python = "0.21.0"
tree-sitter-java = "0.21.0"
rapidfuzz = "^3.9.5"
gitpython = "^3.1.43"
unidiff = "^0.7.5"

numpy = "^1.0"
pylint = "^3.2.6"

# For LLM integrations
litellm = "^1.51.2"
instructor = "<=1.6.3"

openai = "^1.41.0"
anthropic = "^0.40.0"

botocore = "^1.35.54"
boto3 = "^1.35.54"

# Vector Index
llama-index = "^0.10.65"
llama-index-embeddings-openai = "^0.1.11"
llama-index-embeddings-voyageai = "^0.1.4"
llama-index-readers-file = "^0.1.33"
faiss-cpu = "^1.8.0.post1"
tenacity = ">=8.2.0,<9.0.0"

jsonref = "^1.1.0"

# Testbeds
moatless-testbeds = "^0.0.10"

# Visualizations
scipy = "^1.14.0"
pyarrow = "^17.0.0"
pandas = "^2.2.2"
plotly = "^5.24.0"
streamlit = "^1.38.0"
matplotlib = "^3.9.2"
pygraphviz = "^1.14"

[tool.ruff.lint]
select = [ "B", "DTZ", "E", "F", "I", "LOG", "N", "PLE", "SIM", "T20", "UP",]
ignore = [ "E501", "F401", "UP007" ]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
python-dotenv = "^1.0.1"
mypy = "^1.10.0"
ruff = "^0.5.5"
coverage = "^7.6.1"
